feat: Change Delete Userr

+++ BackEnd/FSU.SPORTIDY/FSU.SPORTIDY.Service/Services/UserService.cs
@@ -528,19 +528,44 @@ namespace FSU.SPORTIDY.Service.Services
             if (existUser != null)
             {
                 // update account
-                existUser.FullName = updateUserRequestModel.FullName;
-                existUser.Description = updateUserRequestModel.Description;
-                existUser.Phone = updateUserRequestModel.PhoneNumber;
-                existUser.Birtday = updateUserRequestModel.Birthday;
-                existUser.Address = updateUserRequestModel.Address;
-                existUser.Gender = updateUserRequestModel.Gender;
-                existUser.Avartar = updateUserRequestModel.Avatar;
-
-                bool checkOldPassword = PasswordHelper.VerifyPassword(updateUserRequestModel.Password, existUser.Password);
-                if (checkOldPassword)
+                if(updateUserRequestModel.FullName != null)
                 {
-                    string newPassword = PasswordHelper.HashPassword(updateUserRequestModel.Password);
-                    existUser.Password = newPassword;
+                    existUser.FullName = updateUserRequestModel.FullName;
+                }
+                if(updateUserRequestModel.Description != null)
+                {
+                    existUser.Description = updateUserRequestModel.Description;
+                }
+                if(updateUserRequestModel.PhoneNumber != null)
+                {
+                    existUser.Phone = updateUserRequestModel.PhoneNumber;
+                }
+                if(updateUserRequestModel.Birthday != null)
+                {
+                    existUser.Birtday = updateUserRequestModel.Birthday;
+                }
+                if(updateUserRequestModel.Address != null)
+                {
+                    existUser.Address = updateUserRequestModel.Address;
+                }
+                if(updateUserRequestModel.Gender != null)
+                {
+                    existUser.Gender = updateUserRequestModel.Gender;
+                }
+                if(updateUserRequestModel.Avatar != null)
+                {
+                    existUser.Avartar = updateUserRequestModel.Avatar;
+                }
+                existUser.IsDeleted = updateUserRequestModel.IsDeleted != null ? updateUserRequestModel.IsDeleted : 0;
+               
+                if(!string.IsNullOrEmpty(updateUserRequestModel.Password))
+                {
+                    bool checkOldPassword = PasswordHelper.VerifyPassword(updateUserRequestModel.Password, existUser.Password);
+                    if (checkOldPassword)
+                    {
+                        string newPassword = PasswordHelper.HashPassword(updateUserRequestModel.Password);
+                        existUser.Password = newPassword;
+                    }
                 }
                 var result = await _unitOfWork.UserRepository.UpdateUserAsync(existUser);
                 if (result > 0)
@@ -690,18 +715,21 @@ namespace FSU.SPORTIDY.Service.Services
 
             try
             {
-                // Create a Firebase Storage client
-                var firebaseStorage = new FirebaseStorage(FirebaseConfig.STORAGE_BUCKET);
-                // Parse the image URL to get the file name
-                var fileNameAvatar = EntityDelete.Avartar.Substring(EntityDelete.Avartar.LastIndexOf('/') + 1);
-                fileNameAvatar = fileNameAvatar.Split('?')[0]; // Remove the query parameters
-                var encodedFileAvatar = Path.GetFileName(fileNameAvatar);
-                var fileNameAvatarOfficial = Uri.UnescapeDataString(encodedFileAvatar);
+                if(EntityDelete.Avartar != null && EntityDelete.Avartar.Contains("firebasestorage"))
+                {
+                    // Create a Firebase Storage client
+                    var firebaseStorage = new FirebaseStorage(FirebaseConfig.STORAGE_BUCKET);
+                    // Parse the image URL to get the file name
+                    var fileNameAvatar = EntityDelete.Avartar.Substring(EntityDelete.Avartar.LastIndexOf('/') + 1);
+                    fileNameAvatar = fileNameAvatar.Split('?')[0]; // Remove the query parameters
+                    var encodedFileAvatar = Path.GetFileName(fileNameAvatar);
+                    var fileNameAvatarOfficial = Uri.UnescapeDataString(encodedFileAvatar);
 
 
-                // Delete the avaatar club from Firebase Storage
-                var fileRefAvatar = firebaseStorage.Child(fileNameAvatarOfficial);
-                await fileRefAvatar.DeleteAsync();
+                    // Delete the avaatar club from Firebase Storage
+                    var fileRefAvatar = firebaseStorage.Child(fileNameAvatarOfficial);
+                    await fileRefAvatar.DeleteAsync();
+                }
 
             }
             catch (Exception ex)
